Map Building and Importing
learned about using assets that use textures, changing the texture AND REIMPORTING changes all of those assets that use that texture. like the colormap for the asset pack I use for the graveyard.

Was having trouble coloring faces. It doesn't exist, godot just can't do that without it being separate materials properly imported.

Also .GLB is a GLTF filetype just so you know

learned that  the color maps imported in a folder are not taken, ONLY THE ROOT, highest folder with the colormap will apply to the objects, any others just avoid errors