DONE DID IT PLAYER POV AND MOVEMENT DONE
did the headbob fix for both stopping moving and going back into running, would snap to the headbob position (it takes some explanation, but simply, its a set position on a track that the camera would be offset to) so now everything is smooth and I think I'm in love now. accidentally implemented what I really wanted for the smooth movement, headbob starting slow and continuing up to the normal (max) amount the longer and faster you move around.