THEE AIM HAS BEEN CONQUERED
I was trying to wrap my head around raycasts for the longest time and how to deal with the dumb way it works.

I have a point the guns aim to right?

I also get a point that the raycast shooting from the player camera hits right?

well for some reason way the raycast works is when it is not being updated, it stays in that same position (but still moving in the global position because it was local to my player).
This then made it really hard to use if statements to: Check if it the position was not changing and to Check if the direction wasn't changing

I then tried

I then tried calculating the distance and when the raycast was bugging out, I'd just switch to setting the aim point to a place ahead of the camera by the distance value.

This still didn't work as it was jittery and uncool

THEN I FINALLY FOUND IT.
literally all I had to do was put an if statement, if raycast.is_colliding(): do the thing

then when its not colliding, the position (relative to the player) does not change, exactly what I want.

I think the main reason I did not think about this was because I was thinking in terms of global_position for some reason.