I have been struggling making a sword
I have to make a pattern on the inside of the sword. I tried using sculpting to do it but I think im just gonna make flat cylinders, chop em in half, and attach em to the sword mesh. This also helps make them not annoyingly high in face count